requirements2.txt
3to2==1.1.1
aiml==0.8.6
alembic==0.9.6
appnope==0.1.0
APScheduler==3.3.1
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
autocorrect==0.3.0
Automat==0.6.0
Babel==2.5.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
banal==0.3.0
bcrypt==3.1.3
beanstalkc==0.4.0
beautifulsoup4==4.5.1
bleach==1.5.0
bokeh==0.12.3
boto3==1.4.7
botocore==1.7.22
CacheMan==2.0.6
certifi==2017.7.27.1
cffi==1.11.3
channels==1.1.8
chardet==3.0.4
click==5.1
cloudpickle==0.4.0
coconut==1.2.2
colorama==0.3.3
configparser==3.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
cryptography==2.1.4
cycler==0.10.0
CyHunspell==1.2.1
daphne==1.4.2
dash==0.19.0
dash-core-components==0.14.0
dash-html-components==0.8.0
dash-renderer==0.11.0
dask==0.15.2
dataset==1.0.3
decorator==4.1.2
diff-match-patch==20121119
diskcache==3.0.1
distributed==1.18.3
Django==1.11.7
django-admin-tools==0.8.1
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-autoslug==1.9.3
django-bootstrap-datepicker==1.2.3
django-bootstrap-form==3.3
django-bootstrap-static==3.3.7.1
django-bootstrap3==9.1.0
django-bootstrap3-datetimepicker==2.2.3
django-bower==5.2.0
django-components==0.1
django-contrib-comments==1.7.3
django-cors-middleware==1.3.1
django-dash==0.5.3
django-dashing==0.3.3
django-datetimepicker==3.14
django-debug-toolbar==1.4
django-dynamic-raw-id==2.3
django-easy-select2==1.3.4
django-elasticsearch-dsl==0.4.4
django-enumfields==0.9.0
django-filter==1.0.4
django-grappelli==2.10.1
django-haystack==2.6.1
django-jet==1.0.6
django-leaflet==0.22.0
django-multiselectfield==0.1.8
django-nine==0.1.13
django-nvd3==0.9.7
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
django-phonenumber-field==2.0.0
django-postgres-extensions==0.9.2
django-pyodbc-azure==1.11.0.0
django-registration-redux==1.8
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-searchable-select==1.5.0
django-seed==0.1.7
django-select-multiple-field==0.4.2
django-static-jquery==2.1.4
django-suit==0.2.25
django-tabular-permissions==1.1.1
django-tastypie==0.14.0
django-tastypie-oauth==0.0.3
django-tastypie-swagger==0.1.4
django-tinymce==2.6.0
django-user-accounts==1.3.1
django-webpack-loader==0.5.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
dtw==1.2
easy-thumbnails==2.4.1
ecdsa==0.13
elasticsearch==5.5.1
elasticsearch-dsl==5.3.0
entrypoints==0.2.2
enum34==1.1.6
Fabric==1.14.0
factory-boy==2.7.0
fake-factory==0.7.2
Faker==0.8.8
feedparser==5.2.1
filebrowser-safe==0.4.6
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Script==2.0.5
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
geopy==1.11.0
google-api-python-client==1.4.1
grappelli-safe==0.4.5
h2==3.0.1
h5py==2.7.1
hashids==1.2.0
HeapDict==1.0.0
hpack==3.0.0
html5lib==0.9999999
httplib2==0.10.3
huey==1.5.6
hyperframe==5.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
ipdb==0.8
ipykernel==4.5.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==5.2.2
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
jsonfield==2.0.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.3.0
jws==0.1.3
jwt==0.5.2
language-check==1.1
language-tool==0.3
LatLon==1.0.2
locket==0.2.0
lorem-ipsum-generator==0.3
Mako==1.0.7
Markdown==2.6.7
MarkupSafe==1.0
matplotlib==1.5.3
Mezzanine==4.2.2
minpubsub==0.2.1
mistune==0.7.3
msgpack-python==0.4.8
Naked==0.1.31
nbconvert==4.2.0
nbformat==4.4.0
ndg-httpsclient==0.4.2
normality==0.5.3
notebook==4.2.3
np==1.0.2
numexpr==2.6.2
numpy==1.11.2
oauth2client==4.1.2
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
ordereddict==1.1
paho-mqtt==1.3.0
pandas==0.22.0
paramiko==2.4.0
partd==0.3.8
pathlib2==2.3.0
patsy==0.4.1
pexpect==4.2.1
phonenumberslite==8.8.10
pickleshare==0.7.4
pif==0.8.2
Pillow==4.3.0
pinax-cli==1.0.0
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.10.1
pinax-webanalytics==2.0.4
plotly==2.1.0
pluggy==0.3.1
priority==1.3.0
prompt-toolkit==1.0.15
psutil==2.1.0
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
PyNaCl==1.2.0
pyodbc==4.0.21
pyOpenSSL==17.5.0
pyparsing==2.1.10
pyproj==1.9.5.1
pytest==3.0.2
pytest-cov==2.2.1
pytest-django==2.9.1
python-crontab==2.2.8
python-dateutil==2.6.1
python-editor==1.0.3
python-jwt==2.0.1
python-mimeparse==1.6.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.3
pyxley==0.1.0
PyYAML==3.12
pyzmq==16.0.1
qtconsole==4.2.1
radar==0.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.7.0
rivescript==1.14.9
rsa==3.4.2
s3fs==0.1.2
s3transfer==0.1.11
scandir==1.6
scikit-learn==0.19.1
scipy==0.18.1
seaborn==0.7.1
selenium==2.53.6
service-identity==17.0.0
shellescape==3.4.1
shortuuid==0.5.0
simple-timer==0.2
simplegeneric==0.8.1
simplejson==3.12.0
singledispatch==3.4.0.3
six==1.11.0
snakeMQ==1.4
sortedcontainers==1.5.7
South==1.0.2
SQLAlchemy==1.1.15
sqlparse==0.1.9
statsmodels==0.8.0
streamz==0.1.0
suit==2.0.2
tables==3.3.0
tblib==1.3.2
terminado==0.6
text-unidecode==1.1
tinydb==3.5.0
toolz==0.8.2
tornado==4.5.2
tox==2.1.1
tqdm==4.19.5
traitlets==4.3.2
transliterate==1.10
tsfresh==0.11.0
Twisted==17.9.0
txaio==2.8.2
typing==3.5.3.0
tzlocal==1.3
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.21.1
virtualenv==15.1.0
vishap==0.1.5
wcwidth==0.1.7
Werkzeug==0.12.2
what3words==2.2.1
widgetsnbextension==1.2.6
zict==0.1.3
zope.interface==4.4.3
